mcp_server/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Google Cloud & Vertex AI
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.13.0
google-cloud-storage==2.10.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
vertexai==1.50.0

# Google Gemini AI
google-generativeai==0.3.2

# Google ADK (Agent Development Kit) - Optional
# google-adk-core==0.1.0
# google-adk-models==0.1.0
# google-adk-tools==0.1.0
# google-genai==0.3.0

# APIs & Data
google-search-results==2.4.2
google-api-python-client==2.107.0
requests==2.31.0
aiohttp==3.9.1
feedparser==6.0.10

# Data Processing
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2

# Media Processing
pillow==10.1.0

# Caching & Database
redis==5.0.1
cachetools==5.3.2

# Utilities
python-multipart==0.0.6
pyyaml==6.0.1
